requirements.txt
google-api-python-client
pandas
google-cloud-storage
google-auth
gcsfs
deap
markdown
streamlit
IPython
pydata_google_auth
google-auth-httplib2
google-auth-oauthlib
python-mimeparse
uritemplate